+++ page/hapus-status.php
@@ -7,9 +7,25 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         if(isset($_GET['id_status'])) {
             try {
                 $id = $_GET["id_status"];
-    
-                $sql = "DELETE FROM status_pegawai WHERE id_status_p = '$id';";
-                $res = $db->query($sql);
+                $cek = "SELECT id_status_p FROM pegawai WHERE id_status_p = '$id';";
+                $res1 = $db->query($cek);
+                    if ($res1) {
+                        if ($res1->num_rows == 1) {
+                            $arr = array();
+                            $arr['status'] = "error";
+                            $arr['message'] = "Tidak dapat menghapus data. Pastikan Data tidak dipakai.";
+                            $response = json_encode($arr);
+                            echo $response;
+                        } else {                               
+                            $sql = "DELETE FROM status_pegawai WHERE id_status_p = '$id';";
+                            $res = $db->query($sql);
+                            $arr = array();
+                            $arr['status'] = "success";
+                            $arr['message'] = "Data berhasil dihapus";
+                            $response = json_encode($arr);
+                            echo $response;
+                        }
+                    }
                 //ini return response tadi
             } catch (\Throwable $th) {
                 //ini mengembalikan semua kemungkinan error
